% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/privates.R
\name{.comb}
\alias{.comb}
\title{Villeneuve's Combination Function}
\usage{
.comb(x, ...)
}
\description{
This function combines the outputs produced on the parallel loop of the Villeneuve and
   Desaulnier's algorithm implementations. It cannot be used for other purposes.
}
\details{
Private function that cannot be used by the package's end-users.
}
\seealso{
Other Private Functions: \code{\link{.get_arc_attributes}}
}
\concept{Private Functions}
